In Astro::UTDF:
Overload raw_record() to be a mutator if it has an argument. As a
mutator it unpacks its argument and populates the attributes that map to
the UTDF record. As an accessor it assembles a UTDF record from the
relevant attributes.

Add hex_record() as a wrapper to raw_record, taking (as a mutator) or
returning (as an accessor) a hexified version of the raw UTDF record.

Add methods new() and clone().

Recode slurp() to use new( raw_record => $buffer, prior_record => ... );

Have all potential mutators croak if passed an argument.

Move doppler_shift() to its proper place in the code.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-UTDF/trunk@1644 f6fd3633-1913-47c2-bad4-372d75c948ec